Trait objc2::declare::InnerIvarType

pub unsafe trait InnerIvarType: Sealed + Encode {
    type Output;
}

Types that may be used in ivars.

This may be either:

IvarBool.

IvarDrop<T>.

IvarEncode<T>.

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!

Safety

You cannot rely on any safety guarantees from this.

Required Associated Types

type Output

The type that an Ivar containing this will dereference to.

E.g. Ivar<IvarDrop<Box<u8>>> will deref to Box<u8>.

Object Safety

This trait is not object safe.