Trait objc2::declare::MethodImplementation

pub trait MethodImplementation: Sealed {
    type Callee: RefEncode + ?Sized;
    type Ret: EncodeReturn;
    type Args: EncodeArguments;
}

Types that can be used as the implementation of an Objective-C method.

This is a sealed trait that is implemented for a lot of extern "C" function pointer types.

Required Associated Types

type Callee: RefEncode + ?Sized

The callee type of the method.

type Ret: EncodeReturn

The return type of the method.

type Args: EncodeArguments

The argument types of the method.