Module objc2::encode::__unstable

Unstable encoding traits.

The traits in this module are unstable, meaning that they are outside the usual SemVer guarantee, and may freely change or be removed in patch versions.

They are provided here mostly for documentation purposes (e.g. you can check EncodeReturn to see which types are possible as the return value of a function pointer).

If you would like to use some of these, please open an issue, then we can discuss making them stable.

Traits

EncodeArguments

Types that represent an ordered group of function arguments, where each argument has an Objective-C type-encoding, or can be converted from one.

EncodeConvertArgument

Represents types that can easily be converted to/from an Encode type.

EncodeConvertReturn

Same as EncodeConvertArgument, but for return types.

EncodeReturn

Types that are safe as the return value from Objective-C.