Trait objc2::encode::__unstable::EncodeArguments

pub trait EncodeArguments: Sealed {
    const ENCODINGS: &'static [Encoding];
}

Types that represent an ordered group of function arguments, where each argument has an Objective-C type-encoding, or can be converted from one.

This is implemented for tuples of up to 16 arguments, where each argument implements EncodeConvertArgument. It is primarily used to make generic code a bit easier.

Note that tuples themselves don’t implement Encode directly, because they’re not FFI-safe!

Required Associated Constants

const ENCODINGS: &'static [Encoding]

The encodings for the arguments.

Object Safety

This trait is not object safe.