Trait objc2::encode::EncodeReturn

pub unsafe trait EncodeReturn: Sealed {
    const ENCODING_RETURN: Encoding;
}

Types that are safe as the return value from Objective-C.

This is a sealed trait, and should not need to be implemented manually.

Safety

Similar to Encode, except the value is only guaranteed to be valid as a return value, both from functions/methods you’re calling, and from declared functions/methods.

It does not have to be valid as e.g. an instance variable, or as an argument to a function.

Required Associated Constants

const ENCODING_RETURN: Encoding

The Objective-C type-encoding for this type.

Object Safety

This trait is not object safe.