Module objc2::foundation

Bindings to the Foundation framework.

This is the std equivalent for Objective-C, containing essential data types, collections, and operating-system services.

See Apple’s documentation.

Philosophy

The Foundation framework is huge! If we aspired to map every API it exposes (a lot of it is just helper methods to make Objective-C more ergonomic), this library would never be finished. Instead, our focus lies on conversion methods, to allow easily using them from Rust.

If you find some API that an object doesn’t expose (but should), we gladly accept pull requests. If it is something that is out of scope, these objects implement the Message trait, so you can always just manually call a method on them using the msg_send! family of macros.

Use of Deref

objc2::foundation uses the Deref trait in a bit special way: All objects deref to their superclasses. For example, NSMutableArray derefs to NSArray, which in turn derefs to NSObject.

Note that this is explicitly recommended against in the documentation and the Rust Design patterns book (see those links for details).

Due to Objective-C objects only ever being accessible behind pointers in the first place, the problems stated there are less severe, and having the implementation just means that everything is much nicer when you actually want to use the objects!

All objects also implement AsRef and AsMut to their superclass, and can be used in Id::into_super, so if you favour explicit conversion, that is a possibility too.

Re-exports

pub use crate::ffi::NSInteger;

pub use crate::ffi::NSUInteger;

Structs

CGPoint

A point in a two-dimensional coordinate system.

CGRect

The location and dimensions of a rectangle.

CGSize

A two-dimensional size.

MainThreadMarker

A marker type taken by functions that can only be executed on the main thread.

NSArray

An immutable ordered collection of objects.

NSAttributedString

A string that has associated attributes for portions of its text.

NSBundle

A representation of the code and resources stored in a bundle directory on disk.

NSData

A static byte buffer in memory.

NSDictionary

NSEnumerator

NSError

Information about an error condition including a domain, a domain-specific error code, and application-specific information.

NSException

A special condition that interrupts the normal flow of program execution.

NSFastEnumerator

NSMutableArray

A growable ordered collection of objects.

NSMutableAttributedString

A mutable string that has associated attributes.

NSMutableData

A dynamic byte buffer in memory.

NSMutableDictionary

A mutable collection of objects associated with unique keys.

NSMutableSet

A growable unordered collection of unique objects.

NSMutableString

A dynamic plain-text Unicode string object.

NSNumber

An object wrapper for primitive scalars.

NSObject

NSProcessInfo

A collection of information about the current process.

NSRange

TODO.

NSSet

An immutable unordered collection of unique objects.

NSString

An immutable, plain-text Unicode string object.

NSThread

A thread of execution.

NSUUID

A universally unique value.

NSValue

A container wrapping any encodable type as an Obective-C object.

NSZone

A type used to identify and manage memory zones.

Enums

NSComparisonResult

Constants that indicate sort order.

Constants

NSNotFound

A value indicating that a requested item couldn’t be found or doesn’t exist.

Traits

NSCopying

NSFastEnumeration

NSMutableCopying

TODO

Functions

is_main_thread

Whether the current thread is the main thread.

is_multi_threaded

Whether the application is multithreaded according to Cocoa.

Type Definitions

CGFloat

The basic type for all floating-point values.

NSAttributedStringKey

Attributes that you can apply to text in an attributed string.

NSErrorDomain

NSErrorUserInfoKey

NSPoint

A point in a Cartesian coordinate system.

NSRect

A rectangle.

NSSize

A two-dimensional size.

NSTimeInterval

A number of seconds.