Trait objc2::mutability::IsIdCloneable

pub unsafe trait IsIdCloneable: Sealed { }

Marker trait for classes where Retained::clone / Id::clone is safe.

Since the Foundation collection types (NSArray<T>, NSDictionary<K, V>, …) act as if they store Retaineds, this also makes certain functionality on those types possible.

This is implemented for classes whose ClassType::Mutability is one of:

Root.

Immutable.

ImmutableWithMutableSubclass.

InteriorMutable.

MainThreadOnly.

Safety

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!