objc2::mutability

Trait IsMainThreadOnly

pub unsafe trait IsMainThreadOnly: Sealed { }

Marker trait for classes that are only available on the main thread.

This is implemented for classes whose ClassType::Mutability is one of:

MainThreadOnly.

Since MainThreadOnly types must be !Send and !Sync, if you hold a type that implements this trait, then you’re guaranteed to be on the main thread (and can get a MainThreadMarker using MainThreadMarker::from).

Safety

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!