Trait objc2::mutability::Mutability

pub trait Mutability: Sealed + Sized { }

Marker trait for the different types of mutability a class can have.

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!

Object Safety

This trait is not object safe.