Struct objc2::rc::Id

pub struct Id<T: ?Sized, O: Ownership> { /* private fields */ }

An pointer for Objective-C reference counted objects.

Id strongly references or “retains” the given object T, and “releases” it again when dropped, thereby ensuring it will be deallocated at the right time.

An Id can either be Owned or Shared, represented with the O type parameter.

If owned, it is guaranteed that there are no other references to the object, and the Id can therefore be mutably dereferenced.

If shared, however, it can only be immutably dereferenced because there may be other references to the object, since a shared Id can be cloned to provide exactly that.

An Id<T, Owned> can be safely converted to a Id<T, Shared> using Id::into_shared or From/Into. The opposite is not safely possible, but the unsafe option Id::from_shared is provided.

Option<Id<T, O>> is guaranteed to have the same size as a pointer to the object.

Comparison to std types

Id<T, Owned> can be thought of as the Objective-C equivalent of Box from the standard library: It is a unique pointer to some allocated object, and that means you’re allowed to get a mutable reference to it.

Likewise, Id<T, Shared> is the Objective-C equivalent of Arc: It is a reference-counting pointer that, when cloned, increases the reference count.

Caveats

If the inner type implements Drop, that implementation will not be called, since there is no way to ensure that the Objective-C runtime will do so. If you need to run some code when the object is destroyed, implement the dealloc method instead.

This allows ?Sized types T, but the intention is to only support when T is an extern type (yet unstable).

Examples

use objc2::msg_send_id;
use objc2::runtime::{Class, Object};
use objc2::rc::{Id, Owned, Shared, WeakId};

let cls = Class::get("NSObject").unwrap();
let obj: Id<Object, Owned> = unsafe { msg_send_id![cls, new] };
// obj will be released when it goes out of scope

// share the object so we can clone it
let obj: Id<_, Shared> = obj.into();
let another_ref = obj.clone();
// dropping our other reference will decrement the retain count
drop(another_ref);

let weak = WeakId::new(&obj);
assert!(weak.load().is_some());
// After the object is deallocated, our weak pointer returns none
drop(obj);
assert!(weak.load().is_none());

let mut owned: Id<T, Owned>;
let mut_ref: &mut T = &mut *owned;
// Do something with `&mut T` here

let shared: Id<T, Shared> = owned.into();
let cloned: Id<T, Shared> = shared.clone();
// Do something with `&T` here

Implementations

impl<T: ClassType + 'static, O: Ownership> Id<T, O>where T::Super: 'static,

pub fn into_super(this: Self) -> Id<T::Super, O>

Convert the object into its superclass.

impl<T: Message + ?Sized, O: Ownership> Id<T, O>

pub unsafe fn new(ptr: *mut T) -> Option<Id<T, O>>

Constructs an Id to an object that already has +1 retain count.

This is useful when you have a retain count that has been handed off from somewhere else, usually Objective-C methods like init, alloc, new, copy, or methods with the ns_returns_retained attribute.

Since most of the above methods create new objects, and you therefore hold unique access to the object, you would often set the ownership to be Owned.

But some immutable objects (like NSString) don’t always return unique references, so in those case you would use Shared.

Returns None if the pointer was null.

Safety

The caller must ensure the given object has +1 retain count, and that the object pointer otherwise follows the same safety requirements as in Id::retain.

Example

let cls: &Class;
let obj: &mut Object = unsafe { msg_send![cls, alloc] };
let obj: Id<Object, Owned> = unsafe { Id::new(msg_send![obj, init]).unwrap() };
// Or utilizing `msg_send_id`:
let obj = unsafe { msg_send_id![cls, alloc] };
let obj: Id<Object, Owned> = unsafe { msg_send_id![obj, init] };
// Or in this case simply just:
let obj: Id<Object, Owned> = unsafe { msg_send_id![cls, new] };

let cls = class!(NSString);
// NSString is immutable, so don't create an owned reference to it
let obj: Id<NSString, Shared> = unsafe { msg_send_id![cls, new] };

pub fn as_ptr(this: &Id<T, O>) -> *const T

Returns a raw pointer to the object.

The pointer is valid for at least as long as the Id is held.

See Id::as_mut_ptr for the mutable equivalent.

This is an associated method, and must be called as Id::as_ptr(obj).

impl<T: Message + ?Sized> Id<T, Owned>

pub fn as_mut_ptr(this: &mut Id<T, Owned>) -> *mut T

Returns a raw mutable pointer to the object.

The pointer is valid for at least as long as the Id is held.

See Id::as_ptr for the immutable equivalent.

This is an associated method, and must be called as Id::as_mut_ptr(obj).

impl<T: Message, O: Ownership> Id<T, O>

pub unsafe fn cast<U: Message>(this: Self) -> Id<U, O>

Convert the type of the given object to another.

This is equivalent to a cast between two pointers.

See Id::into_super and ProtocolObject::from_id for safe alternatives.

This is common to do when you know that an object is a subclass of a specific class (e.g. casting an instance of NSString to NSObject is safe because NSString is a subclass of NSObject).

All 'static objects can safely be cast to Object, since that assumes no specific class.

Safety

You must ensure that the object can be reinterpreted as the given type.

If T is not 'static, you must ensure that U ensures that the data contained by T is kept alive for as long as U lives.

Additionally, you must ensure that any safety invariants that the new type has are upheld.

pub unsafe fn retain(ptr: *mut T) -> Option<Id<T, O>>

Retains the given object pointer.

This is useful when you have been given a pointer to an object from some API, and you would like to ensure that the object stays around so that you can work with it.

If said API is a normal Objective-C method, you probably want to use Id::retain_autoreleased instead.

This is rarely used to construct owned Ids, see Id::new for that.

Returns None if the pointer was null.

Safety

The caller must ensure that the ownership is correct; that is, there must be no Owned pointers or mutable references to the same object, and when creating owned Ids, there must be no other pointers or references to the object.

Additionally, the pointer must be valid as a reference (aligned, dereferencable and initialized, see the std::ptr module for more information).

Finally, if you do not know the concrete type of T, it may not be 'static, and hence you must ensure that the data that T references lives for as long as T.

pub unsafe fn retain_autoreleased(ptr: *mut T) -> Option<Id<T, O>>

Retains a previously autoreleased object pointer.

This is useful when calling Objective-C methods that return autoreleased objects, see Cocoa’s Memory Management Policy.

This has exactly the same semantics as Id::retain, except it can sometimes avoid putting the object into the autorelease pool, possibly yielding increased speed and reducing memory pressure.

Note: This relies heavily on being inlined right after msg_send!, be careful not accidentally require instructions between these.

Safety

Same as Id::retain.

pub fn autorelease_return(self) -> *mut T

Autoreleases and prepares the Id to be returned to Objective-C.

The object is not immediately released, but will be when the innermost autorelease pool is drained.

This is useful when declaring your own methods where you will often find yourself in need of returning autoreleased objects to properly follow Cocoa’s Memory Management Policy.

To that end, you could use Id::autorelease, but that would require you to have an AutoreleasePool object at hand, which you clearly won’t have in such cases. This function doesn’t require a pool object (but as a downside returns a pointer instead of a reference).

This is also more efficient than a normal autorelease, it makes a best effort attempt to hand off ownership of the retain count to a subsequent call to objc_retainAutoreleasedReturnValue / Id::retain_autoreleased in the enclosing call frame. Note: This optimization relies heavily on this function being tail called, so be careful to call this function at the end of your method.

Examples

use objc2::{class, msg_send_id, sel};
use objc2::declare::ClassBuilder;
use objc2::rc::{Id, Owned};
use objc2::runtime::{Class, Object, Sel};

let mut builder = ClassBuilder::new("ExampleObject", class!(NSObject)).unwrap();

extern "C" fn get(cls: &Class, _cmd: Sel) -> *mut Object {
    let obj: Id<Object, Owned> = unsafe { msg_send_id![cls, new] };
    obj.autorelease_return()
}

unsafe {
    builder.add_class_method(
        sel!(get),
        get as extern "C" fn(_, _) -> _,
    );
}

let cls = builder.register();

impl<T: Message> Id<T, Owned>

pub fn autorelease<'p>(self, pool: &'p AutoreleasePool) -> &'p mut T

Autoreleases the owned Id, returning a mutable reference bound to the pool.

The object is not immediately released, but will be when the innermost / current autorelease pool (given as a parameter) is drained.

pub unsafe fn from_shared(obj: Id<T, Shared>) -> Self

Promote a shared Id to an owned one, allowing it to be mutated.

Safety

The caller must ensure that there are no other pointers (including WeakId pointers) to the same object.

This also means that the given Id should have a retain count of exactly 1 (except when autoreleases are involved).

In general, this is wildly unsafe, do see if you can find a different solution!

pub fn into_shared(obj: Self) -> Id<T, Shared>

Convert an owned to a shared Id, allowing it to be cloned.

This is also implemented as a From conversion, but this name is more explicit, which may be useful in some cases.

impl<T: Message> Id<T, Shared>

pub fn autorelease<'p>(self, pool: &'p AutoreleasePool) -> &'p T

Autoreleases the shared Id, returning an aliased reference bound to the pool.

The object is not immediately released, but will be when the innermost / current autorelease pool (given as a parameter) is drained.