Trait objc2::rc::DefaultId

pub trait DefaultId {
    type Ownership: Ownership;

    fn default_id() -> Id<Self, Self::Ownership> ;
}

Helper trait to implement Default on types whoose default value is an Id.

Required Associated Types

type Ownership: Ownership

Indicates whether the default value is mutable or immutable.

Required Methods

fn default_id() -> Id<Self, Self::Ownership>

The default Id for a type.

On most objects the implementation would just be sending a message to the new selector.