Trait objc2::rc::IdIntoIterator

pub trait IdIntoIterator {
    type Item;
    type IntoIter: Iterator<Item = Self::Item>;

    // Required method
    fn id_into_iter(this: Retained<Self>) -> Self::IntoIter;

    // Provided method
    fn retained_into_iter(this: Retained<Self>) -> Self::IntoIter { ... }
}

Helper trait to implement IntoIterator on Retained.

This should be implemented in exactly the same fashion as if you were implementing IntoIterator for your type normally.

Required Associated Types

type Item

The type of the elements being iterated over.

type IntoIter: Iterator<Item = Self::Item>

Which kind of iterator are we turning this into?

Required Methods

fn id_into_iter(this: Retained<Self>) -> Self::IntoIter

Creates an iterator from an Retained.

Soft-deprecated alias of RetainedIntoIterator::retained_into_iter.

Provided Methods

fn retained_into_iter(this: Retained<Self>) -> Self::IntoIter

Creates an iterator from an Retained.

You would normally not call this function directly; instead, you’d call into_iter on an Retained.

Object Safety

This trait is not object safe.