Trait objc2::rc::Ownership

pub trait Ownership: Sealed + 'static + Sized + Send + Sync + Unpin + UnwindSafe + RefUnwindSafe + Clone + Copy + PartialEq + Eq + PartialOrd + Ord + Hash + Debug + AutoreleaseSafe { }

A type that marks what type of ownership a struct has over the object(s) it contains; specifically, either Owned or Shared.

This trait is sealed and not meant to be implemented outside of the this crate.

Object Safety

This trait is not object safe.