Struct objc2::runtime::AnyClass

#[repr(C)]
pub struct AnyClass(/* private fields */);

A type that represents an Objective-C class.

This is an opaque type meant to be used behind a shared reference &AnyClass, which is semantically equivalent to Class _Nonnull.

A nullable class can be used as Option<&AnyClass>.

See Apple’s documentation.

Implementations

impl AnyClass

pub fn get(name: &str) -> Option<&'static Self>

Returns the class definition of a specified class, or None if the class is not registered with the Objective-C runtime.

pub fn classes() -> impl Deref<Target = [&'static Self]> + AsRef<[&'static Self]> + Debug

Obtains the list of registered class definitions.

pub fn classes_count() -> usize

Returns the total number of registered classes.

pub fn name(&self) -> &str

Returns the name of the class.

pub fn superclass(&self) -> Option<&AnyClass>

Returns the superclass of self, or None if self is a root class.

pub fn metaclass(&self) -> &Self

Returns the metaclass of self.

pub fn instance_size(&self) -> usize

Returns the size of instances of self.

pub fn instance_method(&self, sel: Sel) -> Option<&Method>

Returns a specified instance method for self, or None if self and its superclasses do not contain an instance method with the specified selector.

pub fn class_method(&self, sel: Sel) -> Option<&Method>

Returns a specified class method for self, or None if self and its superclasses do not contain a class method with the specified selector.

Same as cls.metaclass().class_method().

pub fn instance_variable(&self, name: &str) -> Option<&Ivar>

Returns the ivar for a specified instance variable of self, or None if self has no ivar with the given name.

If the instance variable was not found on the specified class, the superclasses are searched.

Attempting to access or modify instance variables of a class that you do no control may invoke undefined behaviour.

pub fn instance_methods( &self, ) -> impl Deref<Target = [&Method]> + AsRef<[&Method]> + Debug

Describes the instance methods implemented by self.

pub fn conforms_to(&self, proto: &AnyProtocol) -> bool

Checks whether this class conforms to the specified protocol.

pub fn adopted_protocols( &self, ) -> impl Deref<Target = [&AnyProtocol]> + AsRef<[&AnyProtocol]> + Debug

Get a list of the protocols to which this class conforms.

pub fn instance_variables( &self, ) -> impl Deref<Target = [&Ivar]> + AsRef<[&Ivar]> + Debug

Describes the instance variables declared by self.

pub fn responds_to(&self, sel: Sel) -> bool

Check whether instances of this class respond to the given selector.

This doesn’t call respondsToSelector:, but works entirely within the runtime, which means it’ll always be safe to call, but may not return exactly what you’d expect if respondsToSelector: has been overwritten.

That said, it will always return true if an instance of the class responds to the selector, but may return false if they don’t directly (e.g. does so by using forwarding instead).

pub fn verify_sel<A, R>(&self, sel: Sel) -> Result<(), VerificationError>
where A: EncodeArguments, R: EncodeReturn,

Verify argument and return types for a given selector.

This will look up the encoding of the method for the given selector and return a VerificationError if any encodings differ for the arguments A and return type R.

Example

use objc2::{class, sel};
use objc2::runtime::{AnyClass, Bool};
let cls = class!(NSObject);
let sel = sel!(isKindOfClass:);
// Verify that `isKindOfClass:`:
// - Exists on the class
// - Takes a class as a parameter
// - Returns a BOOL
let result = cls.verify_sel::<(&AnyClass,), Bool>(sel);
assert!(result.is_ok());