Struct objc2::runtime::AnyProtocol

#[repr(C)]
pub struct AnyProtocol(/* private fields */);

A type that represents an Objective-C protocol.

Implementations

impl AnyProtocol

pub fn get(name: &str) -> Option<&'static Self>

Returns the protocol definition of a specified protocol, or None if the protocol is not registered with the Objective-C runtime.

pub fn protocols() -> impl Deref<Target = [&'static Self]> + AsRef<[&'static Self]> + Debug

Obtains the list of registered protocol definitions.

pub fn adopted_protocols( &self, ) -> impl Deref<Target = [&AnyProtocol]> + AsRef<[&AnyProtocol]> + Debug

Get a list of the protocols to which this protocol conforms.

pub fn conforms_to(&self, proto: &AnyProtocol) -> bool

Checks whether this protocol conforms to the specified protocol.

pub fn name(&self) -> &str

Returns the name of self.