Struct objc2::runtime::Bool

pub struct Bool { /* private fields */ }

The Objective-C BOOL type.

This is a thin wrapper-type over objc_sys::BOOL. It is intended that you convert this into a Rust bool with the Bool::as_bool method as soon as possible.

This is FFI-safe and can be used directly with msg_send! and extern functions.

Note that this is able to contain more states than bool on some platforms, but these cases should not be relied on!

Implementations

impl Bool

pub const YES: Self = _

The equivalent of true for Objective-C’s BOOL type.

pub const NO: Self = _

The equivalent of false for Objective-C’s BOOL type.

pub const fn new(value: bool) -> Self

Creates an Objective-C boolean from a Rust boolean.

pub const fn from_raw(value: BOOL) -> Self

Creates this from a boolean value received from a raw Objective-C API.

pub const fn as_raw(self) -> BOOL

Retrieves the inner ffi::BOOL boolean type, to be used in raw Objective-C APIs.

pub const fn is_false(self) -> bool

Returns true if self is NO.

You should prefer using as_bool.

pub const fn is_true(self) -> bool

Returns true if self is not NO.

You should prefer using as_bool.

pub const fn as_bool(self) -> bool

Converts this into the bool equivalent.