Struct objc2::runtime::Object

#[repr(C)]
pub struct Object(/* private fields */);

An Objective-C object.

This is slightly different from NSObject in that it may represent an instance of an arbitary Objective-C class (e.g. it does not have to be a subclass of NSObject).

Id<Object, _> is equivalent to Objective-C’s id.

This contains UnsafeCell, and is similar to that in that one can safely access and perform interior mutability on this (both via. msg_send! and through ivars), so long as Rust’s mutability rules are upheld, and that data races are avoided.

Note: This is intentionally neither Sync, Send, UnwindSafe, RefUnwindSafe nor Unpin, since that is something that may change depending on the specific subclass. For example, NSAutoreleasePool is not Send, it has to be deallocated on the same thread that it was created. NSLock is not Send either.

Implementations

impl Object

pub fn class(&self) -> &Class

Dynamically find the class of this object.

pub unsafe fn ivar_ptr<T: Encode>(&self, name: &str) -> *mut T

Returns a pointer to the instance variable / ivar with the given name.

This is similar to UnsafeCell::get, see that for more information on what is and isn’t safe to do.

Usually you will have defined the instance variable yourself with ClassBuilder::add_ivar, the type of the ivar T must match the type used in that.

Attempting to access or modify private implementation details of a class that you do no control using this is not supported, and may invoke undefined behaviour.

Library implementors are strongly encouraged to expose a safe interface to the ivar.

Panics

May panic if the object has no ivar with the given name. May also panic if the type encoding of the ivar differs from the type encoding of T.

This should purely seen as help while debugging and is not guaranteed (e.g. it may be disabled when debug_assertions are off).

Safety

The object must have an instance variable with the given name, and it must be of type T. Any invariants that the object have assumed about the value of the instance variable must not be violated.

No thread syncronization is done on accesses to the variable, so you must ensure that any access to the returned pointer do not cause data races, and that Rust’s mutability rules are not otherwise violated.

pub unsafe fn ivar<T: Encode>(&self, name: &str) -> &T

Returns a reference to the instance variable with the given name.

See Object::ivar_ptr for more information, including on when this panics.

Safety

The object must have an instance variable with the given name, and it must be of type T.

No thread syncronization is done, so you must ensure that no other thread is concurrently mutating the variable. This requirement can be considered upheld if all mutation happens through Object::ivar_mut (since that takes &mut self).

pub unsafe fn get_ivar<T: Encode>(&self, name: &str) -> &T

👎Deprecated: Use Object::ivar instead.

Use Object::ivar instead.

Safety

See Object::ivar.

pub unsafe fn ivar_mut<T: Encode>(&mut self, name: &str) -> &mut T

Returns a mutable reference to the ivar with the given name.

See Object::ivar_ptr for more information, including on when this panics.

Safety

The object must have an instance variable with the given name, and it must be of type T.

This access happens through &mut self, which means we know it to be the only reference, hence you do not need to do any work to ensure that data races do not happen.

pub unsafe fn get_mut_ivar<T: Encode>(&mut self, name: &str) -> &mut T

👎Deprecated: Use Object::ivar_mut instead.

Use Object::ivar_mut instead.

Safety

Same as Object::ivar_mut.

pub unsafe fn set_ivar<T: Encode>(&mut self, name: &str, value: T)

Sets the value of the ivar with the given name.

This is just a helpful shorthand for Object::ivar_mut, see that for more information.

Safety

Same as Object::ivar_mut.