Struct objc2::runtime::Sel

pub struct Sel { /* private fields */ }

A method selector.

The Rust equivalent of Objective-C’s SEL type. You can easily create this using the sel! macro.

The main reason the Objective-C runtime uses a custom types for selectors is to support efficient comparison - a selector is effectively just an interned string, so this makes that very easy!

This guarantees the null-pointer optimization, namely that Option<Sel> is the same size as Sel.

Implementations

impl Sel

pub const fn as_ptr(&self) -> *const objc_selector

Get a pointer to the raw selector.

Useful when working with raw FFI methods.

pub fn register(name: &str) -> Self

Registers a selector with the Objective-C runtime.

This is the dynamic version of the sel! macro, prefer to use that when your selector is static.

Panics

Panics if name contains an internal NUL byte, or if the runtime failed allocating space for the selector.

pub fn name(&self) -> &str

Returns the string representation of the selector.

Panics

Panics if the selector is not valid UTF-8 (however unlikely!)