objc2::runtime

Trait ImplementedBy

pub unsafe trait ImplementedBy<T: ?Sized + Message> { }

An internal helper trait for ProtocolObject.

Safety

This is meant to be a sealed trait, and should not be implemented outside of the extern_protocol! macro.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.