Trait objc2::runtime::MethodImplementation

pub trait MethodImplementation: Sealed + Sized {
    type Callee: ?Sized + RefEncode;
    type Arguments: EncodeArguments;
    type Return: EncodeReturn;
}

Types that can be used as the implementation of an Objective-C method.

This is a sealed trait that is implemented for a lot of extern "C" function pointer types.

Required Associated Types

type Callee: ?Sized + RefEncode

The callee type of the method.

type Arguments: EncodeArguments

The argument types of the method.

type Return: EncodeReturn

The return type of the method.

Object Safety

This trait is not object safe.