objc2::topics::about_generated

Module list

List of framework crates

The following is a full list of all supported Apple frameworks, and the corresponding Rust crate.

Framework	Crate	Docs.rs

ARKit	objc2-ar-kit

AVFAudio	objc2-avf-audio

AVFoundation	objc2-av-foundation

AVKit	objc2-av-kit

AVRouting	objc2-av-routing

Accessibility	objc2-accessibility

AccessorySetupKit	objc2-accessory-setup-kit

Accounts	objc2-accounts

AdServices	objc2-ad-services

AdSupport	objc2-ad-support

AppClip	objc2-app-clip

AppKit	objc2-app-kit

AppTrackingTransparency	objc2-app-tracking-transparency

ApplicationServices	objc2-application-services

AudioToolbox	objc2-audio-toolbox

AuthenticationServices	objc2-authentication-services

AutomaticAssessmentConfiguration	objc2-automatic-assessment-configuration

Automator	objc2-automator

BackgroundAssets	objc2-background-assets

BackgroundTasks	objc2-background-tasks

BrowserEngineCore	objc2-browser-engine-core

BrowserEngineKit	objc2-browser-engine-kit

BusinessChat	objc2-business-chat

CFNetwork	objc2-cf-network

CallKit	objc2-call-kit

CarPlay	objc2-car-play

Carbon	objc2-carbon

Cinematic	objc2-cinematic

ClassKit	objc2-class-kit

ClockKit	objc2-clock-kit

CloudKit	objc2-cloud-kit

Collaboration	objc2-collaboration

ColorSync	objc2-color-sync

CompositorServices	objc2-compositor-services

Contacts	objc2-contacts

ContactsUI	objc2-contacts-ui

CoreAudio	objc2-core-audio

CoreAudioKit	objc2-core-audio-kit

CoreAudioTypes	objc2-core-audio-types

CoreBluetooth	objc2-core-bluetooth

CoreData	objc2-core-data

CoreFoundation	objc2-core-foundation

CoreGraphics	objc2-core-graphics

CoreHaptics	objc2-core-haptics

CoreImage	objc2-core-image

CoreLocation	objc2-core-location

CoreLocationUI	objc2-core-location-ui

CoreMIDI	objc2-core-midi

CoreML	objc2-core-ml

CoreMedia	objc2-core-media

CoreMediaIO	objc2-core-media-io

CoreMotion	objc2-core-motion

CoreNFC	objc2-core-nfc

CoreServices	objc2-core-services

CoreSpotlight	objc2-core-spotlight

CoreTelephony	objc2-core-telephony

CoreText	objc2-core-text

CoreVideo	objc2-core-video

CoreWLAN	objc2-core-wlan

CryptoTokenKit	objc2-crypto-token-kit

DataDetection	objc2-data-detection

DeviceCheck	objc2-device-check

DeviceDiscoveryExtension	objc2-device-discovery-extension

DiskArbitration	objc2-disk-arbitration

EventKit	objc2-event-kit

EventKitUI	objc2-event-kit-ui

ExceptionHandling	objc2-exception-handling

ExecutionPolicy	objc2-execution-policy

ExposureNotification	objc2-exposure-notification

ExtensionKit	objc2-extension-kit

ExternalAccessory	objc2-external-accessory

FSKit	objc2-fs-kit

FileProvider	objc2-file-provider

FileProviderUI	objc2-file-provider-ui

FinderSync	objc2-finder-sync

Foundation	objc2-foundation

GameController	objc2-game-controller

GameKit	objc2-game-kit

GameplayKit	objc2-gameplay-kit

HealthKit	objc2-health-kit

HealthKitUI	objc2-health-kit-ui

HomeKit	objc2-home-kit

IOBluetooth	objc2-io-bluetooth

IOBluetoothUI	objc2-io-bluetooth-ui

IOKit	objc2-io-kit

IOSurface	objc2-io-surface

IOUSBHost	objc2-io-usb-host

IdentityLookup	objc2-identity-lookup

IdentityLookupUI	objc2-identity-lookup-ui

ImageCaptureCore	objc2-image-capture-core

ImageIO	objc2-image-io

InputMethodKit	objc2-input-method-kit

Intents	objc2-intents

IntentsUI	objc2-intents-ui

JavaScriptCore	objc2-javascript-core

LatentSemanticMapping	objc2-latent-semantic-mapping

LinkPresentation	objc2-link-presentation

LocalAuthentication	objc2-local-authentication

LocalAuthenticationEmbeddedUI	objc2-local-authentication-embedded-ui

MLCompute	objc2-ml-compute

MailKit	objc2-mail-kit

MapKit	objc2-map-kit

MediaAccessibility	objc2-media-accessibility

MediaExtension	objc2-media-extension

MediaPlayer	objc2-media-player

MediaSetup	objc2-media-setup

MediaToolbox	objc2-media-toolbox

MessageUI	objc2-message-ui

Messages	objc2-messages

Metal	objc2-metal

MetalFX	objc2-metal-fx

MetalKit	objc2-metal-kit

MetalPerformanceShaders	objc2-metal-performance-shaders

MetalPerformanceShadersGraph	objc2-metal-performance-shaders-graph

MetricKit	objc2-metric-kit

ModelIO	objc2-model-io

MultipeerConnectivity	objc2-multipeer-connectivity

NaturalLanguage	objc2-natural-language

NearbyInteraction	objc2-nearby-interaction

NetworkExtension	objc2-network-extension

NotificationCenter	objc2-notification-center

OSAKit	objc2-osa-kit

OSLog	objc2-os-log

OpenDirectory	objc2-open-directory

PDFKit	objc2-pdf-kit

PHASE	objc2-phase

ParavirtualizedGraphics	objc2-paravirtualized-graphics

PassKit	objc2-pass-kit

PencilKit	objc2-pencil-kit

Photos	objc2-photos

PhotosUI	objc2-photos-ui

PreferencePanes	objc2-preference-panes

PushKit	objc2-push-kit

PushToTalk	objc2-push-to-talk

Quartz	objc2-quartz

QuartzCore	objc2-quartz-core

QuickLook	objc2-quick-look

QuickLookThumbnailing	objc2-quick-look-thumbnailing

QuickLookUI	objc2-quick-look-ui

ReplayKit	objc2-replay-kit

SafariServices	objc2-safari-services

SafetyKit	objc2-safety-kit

SceneKit	objc2-scene-kit

ScreenCaptureKit	objc2-screen-capture-kit

ScreenSaver	objc2-screen-saver

ScreenTime	objc2-screen-time

ScriptingBridge	objc2-scripting-bridge

Security	objc2-security

SecurityFoundation	objc2-security-foundation

SecurityInterface	objc2-security-interface

SensitiveContentAnalysis	objc2-sensitive-content-analysis

SensorKit	objc2-sensor-kit

ServiceManagement	objc2-service-management

SharedWithYou	objc2-shared-with-you

SharedWithYouCore	objc2-shared-with-you-core

ShazamKit	objc2-shazam-kit

Social	objc2-social

SoundAnalysis	objc2-sound-analysis

Speech	objc2-speech

SpriteKit	objc2-sprite-kit

StoreKit	objc2-store-kit

Symbols	objc2-symbols

SystemConfiguration	objc2-system-configuration

SystemExtensions	objc2-system-extensions

TVMLKit	objc2-tv-ml-kit

TVServices	objc2-tv-services

TVUIKit	objc2-tv-ui-kit

ThreadNetwork	objc2-thread-network

UIKit	objc2-ui-kit

UniformTypeIdentifiers	objc2-uniform-type-identifiers

UserNotifications	objc2-user-notifications

UserNotificationsUI	objc2-user-notifications-ui

VideoSubscriberAccount	objc2-video-subscriber-account

VideoToolbox	objc2-video-toolbox

Virtualization	objc2-virtualization

Vision	objc2-vision

WatchConnectivity	objc2-watch-connectivity

WatchKit	objc2-watch-kit

WebKit	objc2-web-kit

iTunesLibrary	objc2-itunes-library

Unsupported

Unsupported frameworks are listed below, feel free to open an issue if a framework that you need isn’t supported, or if you disagree with the given reasoning.

Framework	Why is this unsupported?

AGL	OpenGL-specific, use Metal instead.

Accelerate	Very C-centric, hard for us to map.

ActivityKit	Swift-only.

AdAttributionKit	Swift-only.

AddressBook	Deprecated, use Contacts instead.

AddressBookUI	Deprecated, use Contacts instead.

AppIntents	Swift-only.

AppleScriptKit	Basically empty nowadays.

AppleScriptObjC	Basically empty nowadays.

AssetsLibrary	Deprecated, use PhotoKit instead.

Assignables	Swift-only.

AudioUnit	Deprecated, use AudioToolbox instead.

AudioVideoBridging	Deprecated, use AVKit/AVFoundation instead (maybe?).

AutomatedDeviceEnrollment	Swift-only.

BrowserKit	TODO.

CalendarStore	Very deprecated.

CarKey	Swift-only.

Charts	Swift-only.

Cocoa	Basically empty nowadays, use AppKit.

Combine	Swift-only.

ContactProvider	Mostly Swift-only.

CoreDisplay	Basically empty.

CoreHID	Swift-only.

CoreMIDIServer	Very deprecated.

CoreTransferable	Swift-only.

CreateML	Swift-only.

CreateMLComponents	Swift-only.

CryptoKit	Swift-only.

DVDPlayback	Deprecated, use AVKit/AVFoundation instead.

DeveloperToolsSupport	Swift-only.

DeviceActivity	Swift-only.

DeviceDiscoveryUI	Needs Network first.

DirectoryService	Deprecated, use OpenDirectory instead.

DiscRecording	Deprecated, use AVKit/AVFoundation instead.

DiscRecordingUI	Deprecated, use AVKit/AVFoundation instead.

DockKit	Swift-only.

DriverKit	Uses C++ classes.

ExtensionFoundation	Swift-only.

FamilyControls	Swift-only.

FinanceKit	Swift-only.

FinanceKitUI	Swift-only.

ForceFeedback	Very C-centric and old.

GLKit	OpenGL-specific, use Metal instead.

GLUT	OpenGL-specific, use Metal instead.

GSS	Very C-centric and old.

GroupActivities	Swift-only.

Hypervisor	Very low-level, consider crates like applevisor instead.

ICADevices	Deprecated, use ImageCaptureCore instead.

ImagePlayground	Swift-only.

InstallerPlugins	Deprecated.

InstantMessage	Deprecated in macOS 10.9.

JavaNativeFoundation	Basically empty.

JavaRuntimeSupport	Probably not really interesting..

JournalingSuggestions	Swift-only.

Kerberos	Too low-level.

Kernel	Too low-level.

KernelManagement	Basically empty.

LDAP	Basically empty.

LightweightCodeRequirements	Swift-only.

LiveCommunicationKit	Swift-only.

LockedCameraCapture	Swift-only.

ManagedApp	Swift-only.

ManagedAppDistribution	Swift-only.

ManagedSettings	Swift-only.

ManagedSettingsUI	Swift-only.

MarketplaceKit	Swift-only.

Matter	Mostly available here.

MatterSupport	Swift-only.

MediaLibrary	Deprecated, use PhotoKit instead.

Message	Basically empty.

MobileCoreServices	Deprecated, use CoreServices + UniformTypeIdentifiers instead.

MusicKit	Swift-only.

NetFS	Deprecated, use macFUSE or FSKit instead (probably).

Network	TODO, see #646.

OpenAL	Very C-centric, use newer Audio frameworks instead.

OpenCL	Very C-centric and old.

OpenGL	OpenGL-specific, use Metal instead.

OpenGLES	OpenGL-specific, use Metal instead.

PCSC	Too low-level, consider crates like pcsc instead.

ProximityReader	Swift-only.

ProximityReaderStub	Basically empty.

QTKit	No headers present in Xcode’s SDK.

RealityFoundation	Swift-only.

RealityKit	Swift-only.

RoomPlan	Swift-only.

Ruby	Very C-centric and old.

SecureElementCredential	Swift-only.

SecurityUI	TODO.

StickerFoundation	Basically empty.

StickerKit	Basically empty.

SwiftData	Swift-only.

SwiftUI	Swift-only.

SwiftUICore	Swift-only.

SyncServices	Deprecated.

System	Deprecated wrapper over libSystem.dylib.

TWAIN	Very C-centric and old.

TabletopKit	Swift-only.

TabularData	Swift-only.

Tcl	Very C-centric and old.

TipKit	Swift-only.

Tk	Very C-centric and old.

Translation	Swift-only.

TranslationUIProvider	Swift-only.

Twitter	Deprecated, use Social instead.

VideoDecodeAcceleration	Very C-centric and old.

VisionKit	Swift-only.

WeatherKit	Swift-only.

WidgetKit	Mostly Swift-only.

WorkoutKit	Swift-only.

iAd	Disabled on server side, use AdServices instead.

vecLib	Very C-centric and old.

vmnet	Very C-centric and old.