objc2::topics

Module cross_compiling

Cross-compiling from Linux/Windows

When compiling for e.g. iOS, you’re always doing cross-compilation, since the host platform (macOS) is different from the target platform (iOS). This means that Apple’s tooling generally has pretty good support for cross-compilation.

To cross-compile Rust applications, you need a few things:

A cross-compiler.

Rust is a cross-compiler by default, though you’ll need to install the standard library for the relevant target via. rustup target add $TARGET, or use Cargo’s unstable -Zbuild-std.

A cross-linker.

ld64 is Apple’s native linker. It used to be open source, though the new implementation hasn’t been open sourced.

LLVM’s lld is a fairly good alternative, it comes bundled with Rust as rust-lld.

Linker stubs (*.tbd files).

Used to figure out which dynamic system library each symbol comes from.

These stubs are located inside the SDK.

If compiling C code using e.g. cc-rs, you will need:

A C cross-compiler (Clang is a good choice here).

C headers. Also located inside the SDK.

In the end, you should end up with something like the following:

$ rustup target add aarch64-apple-darwin
$ export CARGO_TARGET_AARCH64_APPLE_DARWIN_LINKER=rust-lld
$ export SDKROOT=$(pwd)/MacOSX.sdk
$ export CC=clang CXX=clang++ AR=llvm-ar # If compiling C code
$ cargo build --target aarch64-apple-darwin

For more details, see Rust Cross, this has a good explanation of how cross-compiling Rust works in general. OSXCross is also a good resource, it outlines how to create a complete compiler toolchain (including ld64).

SDK

The required MacOSX.sdk, iPhoneOS.sdk etc. are bundled with Xcode. Please make sure to review the Xcode SLA, especially the sections regarding required use on Apple-branded computers.

Once you’ve done that, you can download Xcode here (requires logging in with an Apple ID). If you only need to cross-compile to macOS, downloading just the Command Line Tools for Xcode will suffice.

Once downloaded, you need to extract and find the MacOSX.sdk folder, and pass that to rustc using the SDKROOT environment variable.

It may also be possible to find these SDKs elsewhere on the web by searching for e.g. “MacOSX.sdk”, though we will not endorse any such links here, as the legality of such a re-distribution is a grey area.