Trait objc2::MessageArguments

pub unsafe trait MessageArguments: EncodeArguments { }

Types that may be used as the arguments of an Objective-C message.

This is implemented for tuples of up to 12 arguments, where each argument implements Encode (or can be converted from one).

Safety

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!