Crate objc2

Objective-C interface and runtime bindings

Objective-C was the standard programming language on Apple platforms like macOS, iOS, iPadOS, tvOS and watchOS. It is an object-oriented language centered around “sending messages” to its instances - this can for the most part be viewed as a simple method call.

It has since been superseded by Swift, but most of the core libraries and frameworks that are in use on Apple systems are still written in Objective-C, and hence we would like the ability to interract with these using Rust. This crate enables you to do that, in as safe a manner as possible.

See the document on “Layered Safety” for a bit of an introduction to how the safety in this crate works, and see icrate for higher-level bindings to Apple’s frameworks.

Basic usage

This example illustrates major parts of the functionality in this crate:

First, we allocate a new NSObject using ClassType::alloc. Next, we initialize this object. It is ensured to be deallocated using rc::Id. Now we’re free to send messages to the object to our hearts desire using the msg_send! or msg_send_id! macros (depending on the return type of the method). Finally, the Id goes out of scope, and the object is released and deallocated.

use objc2::{msg_send, msg_send_id, ClassType};
use objc2::ffi::NSUInteger;
use objc2::rc::{Id, Owned, Shared};
use objc2::runtime::{NSObject, NSObjectProtocol};

// Creation

let obj1: Id<NSObject, Owned> = unsafe {
    msg_send_id![NSObject::alloc(), init]
};
// Or we can simply do
let obj2 = NSObject::new();

// Usage

let hash1: NSUInteger = unsafe { msg_send![&obj1, hash] };
let hash2: NSUInteger = unsafe { msg_send![&obj2, hash] };
assert_ne!(hash1, hash2);

let is_kind: bool = unsafe {
    msg_send![&obj1, isKindOfClass: NSObject::class()]
};
assert!(is_kind);

// We're going to create a new reference to the first object, so
// relinquish mutable ownership.
let obj1: Id<NSObject, Shared> = obj1.into();
let obj1_self: Id<NSObject, Shared> = unsafe { msg_send_id![&obj1, self] };
assert_eq!(obj1, obj1_self);

// Deallocation on drop

Note that this very simple example contains a lot of unsafe (which should all ideally be justified with a // SAFETY comment). This is required because our compiler can verify very little about the Objective-C invocation, including all argument and return types used in msg_send!; we could have just as easily accidentally made hash an f32, or any other type, and this would trigger undefined behaviour!

See the icrate crate for much more ergonomic usage of the system frameworks like Foundation, AppKit, UIKit and so on.

Anyhow, all of this unsafe nicely leads us to another feature that this crate has:

Encodings and message type verification

The Objective-C runtime includes encodings for each method that describe the argument and return types. See the encode module for a full overview of what this is.

The important part is: To make message sending safer, all arguments and return values for messages must implement encode::Encode. This allows the Rust compiler to prevent you from passing e.g. a Vec into Objective-C, which would both be UB and leak the vector.

Furthermore, we can take advantage of the encodings provided by the runtime to verify that the types used in Rust actually match the types encoded for the method. This is not a perfect solution for ensuring safety (some Rust types have the same Objective-C encoding, but are not equivalent, such as &T and *const T), but it gets us much closer to it!

When debug_assertions are enabled we check the encoding every time you send a message, and the message send will panic if they are not equivalent.

To take the example above, if we changed the hash method’s return type as in the following example, it’ll panic if debug assertions are enabled:

// Wrong return type - this is UB!
let hash1: f32 = unsafe { msg_send![&obj1, hash] };

This library contains further such debug checks, most of which are enabled by default. To enable all of them, use the "verify" cargo feature.

Crate features

This crate exports several optional cargo features, see Cargo.toml for an overview and description of these.

Support for other Operating Systems

The bindings can be used on Linux or *BSD utilizing the GNUstep Objective-C runtime, see the objc-sys crate for how to configure this.

Other functionality

That was a quick introduction, this library also has support for handling exceptions, the ability to dynamically declare Objective-C classes, advanced reference-counting utilities, and more - peruse the documentation at will!

Re-exports

pub use objc_sys as ffi;

pub use self::encode::Encode;

pub use self::encode::Encoding;

pub use self::encode::RefEncode;

Modules

declare

Functionality for dynamically declaring Objective-C classes.

encode

Support for type-encodings

exception

Objective-C’s @throw and @try/@catch.

rc

Utilities for reference counting Objective-C objects.

runtime

A Rust interface for the functionality of the Objective-C runtime.

Macros

class

Gets a reference to a Class from the given name.

declare_class

Declare a new Objective-C class.

extern_class

Create a new type to represent an Objective-C class.

extern_methods

Define methods on an external class.

extern_protocol

Create a new trait to represent an Objective-C protocol.

msg_send

Send a message to an object or class.

msg_send_boolDeprecated

Deprecated. Use msg_send! instead.

msg_send_id

msg_send! for methods returning id, NSObject*, or similar object pointers.

sel

Register a selector with the Objective-C runtime.

Structs

ProtocolObject

An object representing any object that implements a specified protocol.

VerificationError

Failed verifying selector on a class.

Traits

ClassType

Marks types that represent specific classes.

ImplementedBy

An internal helper trait for ProtocolObject.

Message

Types that can be sent Objective-C messages.

MessageArguments

Types that may be used as the arguments of an Objective-C message.

MessageReceiver

Types that can directly be used as the receiver of Objective-C messages.

ProtocolType

Marks types that represent specific protocols.