Module objc2::encode

Support for type-encodings

This module contains traits for annotating types that has an Objective-C type-encoding: Specifically Encode for structs/numeric types and RefEncode for references.

Additionally, this exports the Encoding and EncodingBox types from objc2-encode, see that crate for a few more details on what Objective-C type-encodings are.

Example

Implementing Encode and RefEncode for a custom type:

use objc2::encode::{Encode, Encoding, RefEncode};

#[repr(C)]
struct MyStruct {
    a: f32, // float
    b: i16, // int16_t
}

unsafe impl Encode for MyStruct {
    const ENCODING: Encoding = Encoding::Struct(
        "MyStruct", // Must use the same name as defined in C header files
        &[
            f32::ENCODING, // Same as Encoding::Float
            i16::ENCODING, // Same as Encoding::Short
        ],
    );
}

// @encode(MyStruct) -> "{MyStruct=fs}"
assert!(MyStruct::ENCODING.equivalent_to_str("{MyStruct=fs}"));

unsafe impl RefEncode for MyStruct {
    const ENCODING_REF: Encoding = Encoding::Pointer(&Self::ENCODING);
}

// @encode(MyStruct*) -> "^{MyStruct=fs}"
assert!(MyStruct::ENCODING_REF.equivalent_to_str("^{MyStruct=fs}"));

See the examples folder for more complex usage.

Caveats

We’ve taken the pragmatic approach with Encode and RefEncode, and have implemented it for as many types as possible (instead of defining a bunch of subtraits for very specific purposes). However, that might sometimes be slightly surprising.

The primary example is (), which doesn’t make sense as a method argument, but is a very common return type, and hence implements Encode.

Modules

__unstable

Unstable encoding traits.

Structs

ParseError

The error that was encountered while parsing an encoding string.

Enums

Encoding

An Objective-C type-encoding.

EncodingBox

The boxed version of Encoding.

Traits

Encode

Types that have an Objective-C type-encoding.

OptionEncode

A helper trait for types that are encodable inside an Option.

RefEncode

Types whoose references has an Objective-C type-encoding.