objc2::encode

Trait EncodeArguments

pub trait EncodeArguments: Sealed {
    const ENCODINGS: &'static [Encoding];
}

Types that represent an ordered group of function arguments, where each argument has an Objective-C type-encoding, or can be converted from one.

This is implemented for tuples of up to 16 arguments, where each argument implements EncodeArgument. It is a sealed trait, and should not need to be implemented manually - it is primarily used to make generic code a bit easier to read and understand.

Note that tuples themselves don’t implement Encode directly, because they’re not FFI-safe!

Required Associated Constants

const ENCODINGS: &'static [Encoding]

The encodings for the arguments.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.