Struct objc2::exception::Exception

pub struct Exception(/* private fields */);

An Objective-C exception.

While highly recommended that any exceptions you intend to throw are subclasses of NSException, this is not required by the runtime (similar to how Rust can panic with arbitary payloads using panic_any).