Macro objc2::declare_class

macro_rules! declare_class {
    {
        $(#[$m:meta])*
        $v:vis struct $name:ident {
            $($fields:tt)*
        }

        $ivar_helper_module_v:vis mod $ivar_helper_module:ident;

        unsafe impl ClassType for $for:ty {
            $(#[inherits($($inheritance_rest:ty),+)])?
            type Super = $superclass:ty;

            const NAME: &'static str = $name_const:literal;
        }

        $($methods:tt)*
    } => { ... };
    {
        $(#[$m:meta])*
        $v:vis struct $name:ident {
            $($fields:tt)*
        }

        unsafe impl ClassType for $for:ty {
            $(#[inherits($($inheritance_rest:ty),+)])?
            type Super = $superclass:ty;

            const NAME: &'static str = $name_const:literal;
        }

        $($methods:tt)*
    } => { ... };
    {
        $(#[$m:meta])*
        $v:vis struct $name:ident;

        unsafe impl ClassType for $for:ty {
            $(#[inherits($($inheritance_rest:ty),+)])?
            type Super = $superclass:ty;

            const NAME: &'static str = $name_const:literal;
        }

        $($methods:tt)*
    } => { ... };
}

Declare a new Objective-C class.

This is mostly just a convenience macro on top of extern_class! and the functionality in the declare module, but it can really help with cutting down on boilerplate, in particular when defining delegate classes!

Specification

This macro consists of three parts:

The class definition + ivar definition + inheritance specification.

A set of method definitions.

A set of protocol definitions.

Class and ivar definition

The class definition works a lot like extern_class!, with the added functionality that you can define custom instance variables on your class, which are then wrapped in a declare::Ivar with the given name, and made accessible through the class. (E.g. you can use self.my_ivar as if it was a normal Rust struct).

The instance variable names are specified as such:

IvarEncode<T, "my_crate_ivar">

IvarBool<"my_crate_ivar">

IvarDrop<T, "my_crate_ivar">

This is special syntax that will be used to generate helper types that implement declare::IvarType, which is then used inside the new struct.

Instance variable names must be unique, and must not conflict with any superclass’ instance variables - this means is is good practice to name them with a prefix of your crate name, or similar.

The class name must be specified in ClassType::NAME, and it must be unique across the entire application. Good practice here is similarly to include your crate name in the prefix.

The class is guaranteed to have been created and registered with the Objective-C runtime after the ClassType::class function has been called.

The macro will generate a dealloc method for you, which will call any Drop impl the class may have.

Method definitions

Within the impl block you can define two types of functions; “associated functions” and “methods”. These are then mapped to the Objective-C equivalents “class methods” and “instance methods”. In particular, if you use self or the special name this (or _this), your method will be registered as an instance method, and if you don’t it will be registered as a class method.

The desired selector can be specified using the #[method(my:selector:)] or #[method_id(my:selector:)] attribute, similar to the extern_methods! macro.

If the #[method_id(...)] attribute is used, the return type must be Option<Id<T, O>> or Id<T, O>. Additionally, if the selector is in the “init”-family, the self/this argument must be Allocated<Self>.

Putting other attributes on the method such as cfg, allow, doc, deprecated and so on is supported. However, note that cfg_attr may not work correctly, due to implementation difficulty - if you have a concrete use-case, please open an issue, then we can discuss it.

A transformation step is performed on the functions (to make them have the correct ABI) and hence they shouldn’t really be called manually. (You can’t mark them as pub for the same reason). Instead, use the extern_methods! macro to create a Rust interface to the methods.

If the argument or return type is bool, a conversion is performed to make it behave similarly to the Objective-C BOOL. Use runtime::Bool if you want to control this manually.

Note that &mut Id<_, _> and other such out parameters are not yet supported, and may generate a panic at runtime.

Protocol definitions

You can specify protocols that the class should implement, along with any required/optional methods for said protocols.

The protocol must have been previously defined with extern_protocol!.

The methods work exactly as normal, they’re only put “under” the protocol definition to make things easier to read.

Putting attributes on the impl item such as cfg, allow, doc, deprecated and so on is supported.

Panics

The implemented ClassType::class method may panic in a few cases, such as if:

A class with the specified name already exists.

One of the class’ instance variables already exist on a superclass.

Debug assertions are enabled, and an overriden method’s signature is not equal to the one on the superclass.

The verify feature and debug assertions are enabled, and the required protocol methods are not implemented.

And possibly more similar cases.

Safety

Using this macro requires writing a few unsafe markers:

unsafe impl ClassType for T has the following safety requirements:

Same as extern_class! (the inheritance chain has to be correct).

Any instance variables you specify under the struct definition must either be able to be created using MaybeUninit::zeroed, or be properly initialized in an init method.

unsafe impl T { ... } asserts that the types match those that are expected when the method is invoked from Objective-C. Note that there are no safe-guards here; you can easily write i8, but if Objective-C thinks it’s an u32, it will cause UB when called!

unsafe impl P for T { ... } requires that all required methods of the specified protocol is implemented, and that any extra requirements (implicit or explicit) that the protocol has are upheld. The methods in this definition has the same safety requirements as above.

Examples

Declare a class MyCustomObject that inherits NSObject, has a few instance variables and methods, and implements the NSCopying protocol.

use std::os::raw::c_int;
use objc2::declare::{Ivar, IvarDrop, IvarEncode};
use objc2::rc::{Id, Owned, Shared};
use objc2::runtime::{NSObject, NSObjectProtocol, NSZone};
use objc2::{
    declare_class, extern_protocol, msg_send, msg_send_id, ClassType,
    ProtocolType,
};

// Declare the NSCopying protocol so that we can implement it.
//
// In practice, you wouldn't have to do this, since it is done for you in
// `icrate`.
extern_protocol!(
    unsafe trait NSCopying {
        #[method(copyWithZone:)]
        fn copy_with_zone(&self, _zone: *const NSZone) -> *mut Self;
    }

    unsafe impl ProtocolType for dyn NSCopying {
        const NAME: &'static str = "NSCopying";
    }
);

declare_class!(
    struct MyCustomObject {
        foo: IvarEncode<u8, "_foo">,
        pub bar: IvarEncode<c_int, "_bar">,
        object: IvarDrop<Id<NSObject, Shared>, "_object">,
    }

    mod ivars;

    unsafe impl ClassType for MyCustomObject {
        type Super = NSObject;
        const NAME: &'static str = "MyCustomObject";
    }

    unsafe impl MyCustomObject {
        #[method(initWithFoo:)]
        fn init_with(this: &mut Self, foo: u8) -> Option<&mut Self> {
            let this: Option<&mut Self> = unsafe {
                msg_send![super(this), init]
            };

            this.map(|this| {
                // Initialize instance variables

                // Some types like `u8`, `bool`, `Option<Box<T>>` and
                // `Option<Id<T, O>>` are safe to zero-initialize, and
                // we can simply write to the variable as normal:
                *this.foo = foo;
                *this.bar = 42;

                // For others like `&u8`, `Box<T>` or `Id<T, O>`, we have
                // to initialize them with `Ivar::write`:
                Ivar::write(&mut this.object, Id::into_shared(NSObject::new()));

                // All the instance variables have been initialized; our
                // initializer is sound
                this
            })
        }

        #[method(foo)]
        fn __get_foo(&self) -> u8 {
            *self.foo
        }

        #[method_id(object)]
        fn __get_object(&self) -> Id<NSObject, Shared> {
            self.object.clone()
        }

        #[method(myClassMethod)]
        fn __my_class_method() -> bool {
            true
        }
    }

    unsafe impl NSCopying for MyCustomObject {
        #[method_id(copyWithZone:)]
        fn copy_with_zone(&self, _zone: *const NSZone) -> Id<Self, Owned> {
            let mut obj = Self::new(*self.foo);
            *obj.bar = *self.bar;
            obj
        }

        // If we have tried to add other methods here, or had forgotten
        // to implement the method, we would have gotten an error with the
        // `verify` feature enabled.
    }
);

impl MyCustomObject {
    pub fn new(foo: u8) -> Id<Self, Owned> {
        unsafe { msg_send_id![Self::alloc(), initWithFoo: foo] }
    }

    pub fn get_foo(&self) -> u8 {
        unsafe { msg_send![self, foo] }
    }

    pub fn get_object(&self) -> Id<NSObject, Shared> {
        unsafe { msg_send_id![self, object] }
    }

    pub fn my_class_method() -> bool {
        unsafe { msg_send![Self::class(), myClassMethod] }
    }
}

// TODO: `NSCopying` from `icrate` works a bit differently
// unsafe impl icrate::Foundation::NSCopying for MyCustomObject {
//     type Ownership = Owned;
//     type Output = Self;
// }

fn main() {
    let obj = MyCustomObject::new(3);
    assert_eq!(*obj.foo, 3);
    assert_eq!(*obj.bar, 42);
    assert!(obj.object.is_kind_of::<NSObject>());

    let obj: Id<MyCustomObject, Shared> = unsafe {
         msg_send_id![&obj, copy]
    }; // Or obj.copy() with `icrate`

    assert_eq!(obj.get_foo(), 3);
    assert!(obj.get_object().is_kind_of::<NSObject>());

    assert!(MyCustomObject::my_class_method());
}

Approximately equivalent to the following ARC-enabled Objective-C code.

#import <Foundation/Foundation.h>

@interface MyCustomObject: NSObject <NSCopying> {
    // Public ivar
    int bar;
}

- (instancetype)initWithFoo:(uint8_t)foo;
- (uint8_t)foo;
- (NSObject*)object;
+ (BOOL)myClassMethod;

@end

@implementation MyCustomObject {
    // Private ivar
    uint8_t foo;
    NSObject* _Nonnull object;
}

- (instancetype)initWithFoo:(uint8_t)foo_arg {
    self = [super init];
    if (self) {
        self->foo = foo_arg;
        self->bar = 42;
        self->object = [NSObject new];
    }
    return self;
}

- (uint8_t)foo {
    return self->foo; // Or just `foo`
}

- (NSObject*)object {
    return self->object;
}

+ (BOOL)myClassMethod {
    return YES;
}

// NSCopying

- (id)copyWithZone:(NSZone *)_zone {
    MyCustomObject* obj = [[MyCustomObject alloc] initWithFoo: self->foo];
    obj->bar = self->bar;
    return obj;
}

@end