objc2::mutability

Trait Mutability

pub trait Mutability: Sealed + Sized { }

Marker trait for the different types of mutability a class can have.

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.