Module objc2::runtime

A Rust interface for the functionality of the Objective-C runtime.

For more information on foreign functions, see Apple’s documentation: https://developer.apple.com/library/mac/documentation/Cocoa/Reference/ObjCRuntimeRef/index.html

Structs

Bool

The Objective-C BOOL type.

Class

A type that represents an Objective-C class.

Ivar

A type that represents an instance variable.

Method

A type that represents a method in a class definition.

NSObject

The root class of most Objective-C class hierarchies.

NSZone

A type used to identify and manage memory zones.

Object

An Objective-C object.

Protocol

A type that represents an Objective-C protocol.

Sel

A method selector.

Constants

NODeprecated

Use Bool::NO or ffi::NO instead.

YESDeprecated

Use Bool::YES or ffi::YES instead.

Traits

NSObjectProtocol

The methods that are fundamental to most Objective-C objects.

Type Aliases

BOOLDeprecated

Use Bool or ffi::BOOL instead.

Imp

A pointer to the start of a method implementation.