Struct objc2::runtime::NSObject

#[repr(C)]
pub struct NSObject { /* private fields */ }

The root class of most Objective-C class hierarchies.

This represents the NSObject class. The name โNSObjectโ also refers to a protocol, see NSObjectProtocol for that.

Since this class is only available with the Foundation framework, objc2 links to it for you.

This is exported under icrate::Foundation::NSObject, you probably want to use that path instead.

Implementations

impl NSObject

pub fn new() -> Id<Self, Owned>

Create a new empty NSObject.