Module objc2::topics::about_generated::list

List of framework crates

The following is a full list of currently supported framework crates.

See #393 if the framework that you need is not here. Note that a lot of frameworks are CoreFoundation-like, those are more difficult to support, see #556 for more information.

Framework	Crate	Documentation

AVKit

Accessibility

Accounts

AdServices

AdSupport

AppKit

AppTrackingTransparency

AuthenticationServices

AutomaticAssessmentConfiguration

Automator

BackgroundAssets

BackgroundTasks

BusinessChat

CallKit

ClassKit

CloudKit

Contacts

ContactsUI

CoreBluetooth

CoreData

CoreImage

CoreLocation

CoreML

CoreMotion

CoreWLAN

DataDetection

DeviceCheck

EventKit

ExceptionHandling

ExtensionKit

ExternalAccessory

FileProvider

FileProviderUI

FinderSync

Foundation

GameController

GameKit

HealthKit

IdentityLookup

InputMethodKit

LinkPresentation

LocalAuthentication

LocalAuthenticationEmbeddedUI

MLCompute

MailKit

MapKit

MediaPlayer

Metal

MetalFX

MetalKit

MetricKit

MultipeerConnectivity

NaturalLanguage

NearbyInteraction

NetworkExtension

OSAKit

Photos

PhotosUI

QuartzCore

ScreenCaptureKit

SensitiveContentAnalysis

ServiceManagement

Social

SoundAnalysis

Speech

StoreKit

Symbols

UIKit

UniformTypeIdentifiers

UserNotifications

Virtualization

Vision

WebKit